refactor: change port parameter type from uint16_t to uint32_t in ANetClient::configure

+++ client/src/networks/interface/ANetClient.cpp
@@ -10,7 +10,7 @@
 namespace Network
 {
 
-    void ANetClient::configure(const std::string &ip, uint16_t port)
+    void ANetClient::configure(const std::string &ip, uint32_t port)
     {
         _ip = ip;
         _port = port;
